Super important update for plugins!
Plugins now keep their variables across calls, so now you can set
variable x to 5 in a command, and it will properly retain its value for
the next time its called such as the command again or an event. For
examples on this check the example plugin; the part of it that says
"this doesn't work" now works.