Added 0/1 permission settings
Added new command setUserPerms to command handler
arg[1] is a username, if arg[2] is "true" then the username is given
permissions to use commands from in game chat.  if not true, permissions
are removed.

The client console is hardcoded to always have full permissions.

Permissions could be improved later by creating more levels of access.

Also commented the use of "self" as default behavior for commands from
client console.