Partial slot switch functionality
multiple changes to invHandler.  now can swapSlots, exchanging their
contents, but fails after the first call.  getItem will return the slot
that a given item is in, and swapTo will change the active slot to
contain a given item.  closeInventory will close the inventory window,
but we never are required to declare when we open it, so that is mainly
for formality, but can also be used to clear the crafting slots of
items.

Use !swap invId invId to swap the contents of two slots.  local
inventory should be updated accordingly.